+++ router/documentation/routes/home-element/examples/index.js
@@ -1,0 +1,7 @@
+context.message = 'hello preview message';
+context.count = 0;
+context.drawer = false;
+
+context.toggleDrawer = () => {
+  context.drawer = !context.drawer;
+};