+++ router/src/update-state-middleware/index.js
@@ -5,7 +5,7 @@
  * @param {Function} next
  */
 export default (request, response, next) => {
-  if (!request.popped) {
+  if (!request.popped && response.matchedRoute !== null) {
     window.history.pushState(null, null, response.to);
   }
   next(response);